@@ -1,18 +1,24 @@
-var application = require("application");
-var googleAnalytics = require("nativescript-google-analytics");
+const application = require("tns-core-modules/application");
+import * as googleAnalytics from "nativescript-google-analytics";
+
 application.cssFile = "./app.css";
 
 if (application.ios) {
     //IOS
     var __extends = this.__extends || function (d, b) {
         for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p];
-        function __() { this.constructor = d; }
+
+        function __() {
+            this.constructor = d;
+        }
+
         __.prototype = b.prototype;
         d.prototype = new __();
     };
 
     var appDelegate = (function (_super) {
         __extends(appDelegate, _super);
+
         function appDelegate() {
             _super.apply(this, arguments);
         }
@@ -25,25 +31,25 @@ if (application.ios) {
         return appDelegate;
     })(UIResponder);
     application.ios.delegate = appDelegate;
-}else{
+} else {
     //ANDROID
-    application.on(application.launchEvent, function (args) {
+    application.on(application.launchEvent, args => {
         initAnalytics(); //Module Code to initalize
     });
 
 }
 
-application.run({ moduleName:"app-root"});
+application.run({moduleName: "app-root"});
 
-function initAnalytics(){
+function initAnalytics() {
     googleAnalytics.initalize({
-                trackingId: "UA-74227193-1",
-                //userId: "9ac7a034-ffde-4783-8374-f78b3df39d32", //Optional
-                enableDemographics: true,
-                dispatchInterval: 5,
-                logging: {
-                    native: true,
-                    console: true
-                }
-            });
+        trackingId: "UA-74227193-1",
+        //userId: "9ac7a034-ffde-4783-8374-f78b3df39d32", //Optional
+        enableDemographics: true,
+        dispatchInterval: 5,
+        logging: {
+            native: true,
+            console: true
+        }
+    });
 }
