@@ -1,12 +1,4 @@
-import { Observable, fromObject, fromObjectRecursive } from "tns-core-modules/data/observable"
-import { ObservableArray } from "tns-core-modules/data/observable-array"
+import {Observable} from "tns-core-modules/data/observable"
 
 export class MainViewModel extends Observable {
-    constructor() {
-        super();
-
-        
-
-    }
-
 }
